requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
discord.py>=2.3.0
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.28.0
plotly>=5.17.0
